# apps/web/openapi/generated/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "CourseLit REST API",
        "version": "1.0.0",
        "description": "OpenAPI documentation for the public CourseLit REST API surface."
    },
    "servers": [
        {
            "url": "/",
            "description": "Current CourseLit school origin"
        }
    ],
...
    "paths": {
...
        "/api/member-mimic": {
            "get": {
                "tags": ["Member Mimic"],
                "operationId": "getMemberMimic",
                "summary": "Verify the current member view",
                "description": "Rechecks tenant, actor session, permission, subject and expiry. Invalid markers return expired, never administrator fallback.",
                "responses": {
                    "200": {
                        "description": "Current view; token is never returned.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "mimic": {
                                            "oneOf": [
                                                {
                                                    "type": "object",
                                                    "required": ["kind"],
                                                    "properties": {
                                                        "kind": {
                                                            "type": "string",
                                                            "enum": ["inactive"]
                                                        }
                                                    }
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "kind",
                                                        "returnTo"
                                                    ],
                                                    "properties": {
                                                        "kind": {
                                                            "type": "string",
                                                            "enum": ["expired"]
                                                        },
                                                        "returnTo": {
                                                            "type": "string"
                                                        }
                                                    }
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "kind",
                                                        "actor",
                                                        "subject",
                                                        "expiresAt",
                                                        "returnTo"
                                                    ],
                                                    "properties": {
                                                        "kind": {
                                                            "type": "string",
                                                            "enum": ["active"]
                                                        },
                                                        "actor": {
                                                            "type": "object",
                                                            "required": [
                                                                "userId",
                                                                "name",
                                                                "email"
                                                            ],
                                                            "properties": {
                                                                "userId": {
                                                                    "type": "string"
                                                                },
                                                                "name": {
                                                                    "type": "string"
                                                                },
                                                                "email": {
                                                                    "type": "string",
                                                                    "format": "email"
                                                                }
                                                            }
                                                        },
                                                        "subject": {
                                                            "type": "object",
                                                            "required": [
                                                                "userId",
                                                                "name",
                                                                "email"
                                                            ],
                                                            "properties": {
                                                                "userId": {
                                                                    "type": "string"
                                                                },
                                                                "name": {
                                                                    "type": "string"
                                                                },
                                                                "email": {
                                                                    "type": "string",
                                                                    "format": "email"
                                                                }
                                                            }
                                                        },
                                                        "expiresAt": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "returnTo": {
                                                            "type": "string"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid or oversized JSON."
                    },
                    "403": {
                        "description": "Use this site's Origin and an authorized administrator session."
                    },
                    "404": {
                        "description": "Site or member is unavailable."
                    },
                    "429": {
                        "description": "Too many starts; retry after the window."
                    },
                    "503": {
                        "description": "Status is unavailable; no success is assumed."
                    }
                }
            },
            "post": {
                "tags": ["Member Mimic"],
                "operationId": "startMemberMimic",
                "summary": "Open the member's normal read-only profile",
                "security": [
                    {
                        "CourseLitSession": []
                    }
                ],
                "description": "Requires user:manage and same-origin JSON. Sets an HttpOnly SameSite Strict cookie and revokes the prior view in the same actor session. Authority expires after 15 minutes; the marker remains for explicit Exit. The body is limited to 2 KB; starts to 12/minute.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "additionalProperties": false,
                                "required": ["userId"],
                                "properties": {
                                    "userId": {
                                        "type": "string",
                                        "minLength": 1,
                                        "maxLength": 128,
                                        "pattern": "^[a-zA-Z0-9_-]+$"
                                    },
                                    "returnTo": {
                                        "type": "string",
                                        "maxLength": 1024,
                                        "description": "Only /dashboard/users plus its filter query is accepted."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "View created and cookie set.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "mimic": {
                                            "oneOf": [
                                                {
                                                    "type": "object",
                                                    "required": ["kind"],
                                                    "properties": {
                                                        "kind": {
                                                            "type": "string",
                                                            "enum": ["inactive"]
                                                        }
                                                    }
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "kind",
                                                        "returnTo"
                                                    ],
                                                    "properties": {
                                                        "kind": {
                                                            "type": "string",
                                                            "enum": ["expired"]
                                                        },
                                                        "returnTo": {
                                                            "type": "string"
                                                        }
                                                    }
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "kind",
                                                        "actor",
                                                        "subject",
                                                        "expiresAt",
                                                        "returnTo"
                                                    ],
                                                    "properties": {
                                                        "kind": {
                                                            "type": "string",
                                                            "enum": ["active"]
                                                        },
                                                        "actor": {
                                                            "type": "object",
                                                            "required": [
                                                                "userId",
                                                                "name",
                                                                "email"
                                                            ],
                                                            "properties": {
                                                                "userId": {
                                                                    "type": "string"
                                                                },
                                                                "name": {
                                                                    "type": "string"
                                                                },
                                                                "email": {
                                                                    "type": "string",
                                                                    "format": "email"
                                                                }
                                                            }
                                                        },
                                                        "subject": {
                                                            "type": "object",
                                                            "required": [
                                                                "userId",
                                                                "name",
                                                                "email"
                                                            ],
                                                            "properties": {
                                                                "userId": {
                                                                    "type": "string"
                                                                },
                                                                "name": {
                                                                    "type": "string"
                                                                },
                                                                "email": {
                                                                    "type": "string",
                                                                    "format": "email"
                                                                }
                                                            }
                                                        },
                                                        "expiresAt": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "returnTo": {
                                                            "type": "string"
                                                        }
                                                    }
                                                }
                                            ]
                                        },
                                        "redirectTo": {
                                            "type": "string",
                                            "enum": ["/dashboard/profile"]
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid or oversized JSON."
                    },
                    "403": {
                        "description": "Use this site's Origin and an authorized administrator session."
                    },
                    "404": {
                        "description": "Site or member is unavailable."
                    },
                    "429": {
                        "description": "Too many starts; retry after the window."
                    },
                    "503": {
                        "description": "Status is unavailable; no success is assumed."
                    }
                }
            },
            "delete": {
                "tags": ["Member Mimic"],
                "operationId": "exitMemberMimic",
                "summary": "Revoke and leave the member view",
                "description": "Requires this site's Origin and application/json. Clears even an expired or malformed marker; does not require a still-active member-view authority.",
                "responses": {
                    "200": {
                        "description": "Server authority revoked and browser cookie cleared.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "redirectTo": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid or oversized JSON."
                    },
                    "403": {
                        "description": "Use this site's Origin and an authorized administrator session."
                    },
                    "404": {
                        "description": "Site or member is unavailable."
                    },
                    "429": {
                        "description": "Too many starts; retry after the window."
                    },
                    "503": {
                        "description": "Status is unavailable; no success is assumed."
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "CourseLitSession": {
                "type": "apiKey",
                "in": "cookie",
                "name": "courselit.session_token",
                "description": "Existing CourseLit browser session (secure cookie may have __Secure- prefix). Writes require same-origin Origin and application/json."
            },
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "x-api-key",
                "description": "API key created in CourseLit dashboard settings. The legacy `apikey` request-body field is still accepted but deprecated."
            }
        }
    }
}